refactor: move cancellableRunScript to run-script.ts, load via @vite-ignore
cancellableRunScript is never called from the renderer — the renderer path
uses window.main.hiddenBrowserWindow.runScript instead. Moving it to a
dedicated Node-only module (run-script.ts) and loading it via

  require(/* @vite-ignore */ './run-script')

prevents Vite from bundling the file into the renderer build, which in turn
removes script-executor.ts (and its require('node:fs/promises')) from the
renderer module graph entirely. The template-literal workaround added
previously is no longer needed and has been reverted.

cancellation.ts now exports cancelRequestFunctionMap so run-script.ts can
share the same cancel map.
