+++ packages/insomnia/src/network/cancellation.ts
@@ -1,8 +1,6 @@
-import type { RequestContext } from '../../../insomnia-scripting-environment/src/objects';
 import type { CurlRequestOptions } from '../main/network/libcurl-promise';
-import { runScript as nodejsRunScript } from '../script-executor';
 
-const cancelRequestFunctionMap = new Map<string, () => void>();
+export const cancelRequestFunctionMap = new Map<string, () => void>();
 
 export async function cancelRequestById(requestId: string) {
   window.main.completeExecutionStep({ requestId });
@@ -34,33 +32,6 @@ export const cancellableExecution = async (options: { id: string; fn: Promise<an
     throw err;
   } finally {
     cancelRequestFunctionMap.delete(options.id);
-  }
-};
-
-export const cancellableRunScript = async (options: { script: string; context: RequestContext }) => {
-  const request = options.context.request;
-  const requestId = request._id;
-  const controller = new AbortController();
-  const cancelRequest = () => {
-    // TODO: implement cancelPreRequestScript on hiddenBrowserWindow side?
-    controller.abort();
-  };
-  cancelRequestFunctionMap.set(requestId, cancelRequest);
-  try {
-    const result = await cancellablePromise({
-      signal: controller.signal,
-      fn: process.type === 'renderer' ? window.main.hiddenBrowserWindow.runScript(options) : nodejsRunScript(options),
-    });
-
-    return result;
-  } catch (err) {
-    if (err.name === 'AbortError') {
-      throw new Error('Request was cancelled');
-    }
-    console.log('[network] Error', err);
-    throw err;
-  } finally {
-    cancelRequestFunctionMap.delete(requestId);
   }
 };
