+++ packages/insomnia/src/network/network.ts
@@ -54,7 +54,6 @@ import { serializeNDJSON } from '../utils/ndjson';
 import { buildQueryStringFromParams, joinUrlAndQueryString, smartEncodeUrl } from '../utils/url/querystring';
 import { QUERY_PARAMS } from './api-key/constants';
 import { getAuthObjectOrNull, isAuthEnabled } from './authentication';
-import { cancellableRunScript } from './cancellation';
 import { filterClientCertificates } from './certificate';
 import { runScriptConcurrently, type TransformedExecuteScriptContext } from './concurrency';
 import { addSetCookiesToToughCookieJar } from './set-cookie-util';
@@ -519,7 +518,10 @@ const tryToExecuteScript = async (context: RequestAndContextAndOptionalResponse)
   }
 
   try {
-    const fn = process.type === 'renderer' ? runScriptConcurrently : cancellableRunScript;
+    const fn =
+      process.type === 'renderer'
+        ? runScriptConcurrently
+        : (require(/* @vite-ignore */ './run-script') as typeof import('./run-script')).cancellableRunScript;
     const output = await fn({
       script,
       context: {
